refactor(sandbox): extract shared args and helpers used by create and fork
