+++ packages/sandbox/src/commands/fork.ts
@@ -7,11 +7,14 @@ import { Duration } from "../types/duration";
 import { scope } from "../args/scope";
 import { sandboxClient } from "../client";
 import { sandboxName } from "../args/sandbox-name";
+import { publishPorts } from "../args/ports";
+import { snapshotRetentionArgs } from "../args/snapshot-retention";
 import * as Exec from "./exec";
 import { networkPolicyArgs } from "../args/network-policy";
 import { buildNetworkPolicy } from "../util/network-policy";
 import { ObjectFromKeyValue } from "../args/key-value-pair";
-import { SnapshotExpiration } from "../types/snapshot-expiration";
+import { buildKeepLastSnapshotsPayload } from "../util/keep-last-snapshots";
+import { printSandboxSummary } from "../util/print-sandbox-summary";
 
 export const args = {
   source: cmd.positional({
@@ -35,28 +38,7 @@ export const args = {
       "Override the maximum sandbox runtime (inherited from source if omitted). Example: 5m, 30m",
   }),
   vcpus,
-  ports: cmd.multioption({
-    long: "publish-port",
-    short: "p",
-    description: "Publish sandbox port(s) to DOMAIN.vercel.run",
-    type: cmd.array(
-      cmd.extendType(cmd.number, {
-        displayName: "PORT",
-        async from(number) {
-          if (number < 1024 || number > 65535) {
-            throw new Error(
-              [
-                `Invalid port: ${number}.`,
-                `${chalk.bold("hint:")} Ports must be between 1024-65535 (privileged ports 0-1023 are reserved).`,
-                "╰▶ Examples: 3000, 8080, 8443",
-              ].join("\n"),
-            );
-          }
-          return number;
-        },
-      }),
-    ),
-  }),
+  ports: publishPorts,
   silent: cmd.flag({
     long: "silent",
     description: "Don't write sandbox name to stdout",
@@ -80,45 +62,7 @@ export const args = {
     description:
       "Key-value tags to associate with the fork (overrides tags copied from the source)",
   }),
-  snapshotExpiration: cmd.option({
-    long: "snapshot-expiration",
-    type: cmd.optional(SnapshotExpiration),
-    description:
-      'Default snapshot expiration. Use "none" or 0 for no expiration. Example: 7d, 30d',
-  }),
-  keepLastSnapshots: cmd.option({
-    long: "keep-last-snapshots",
-    type: cmd.optional(
-      cmd.extendType(cmd.number, {
-        displayName: "COUNT",
-        async from(n) {
-          if (!Number.isInteger(n) || n < 1 || n > 10) {
-            throw new Error(
-              `Invalid --keep-last-snapshots value: ${n}. Must be an integer between 1 and 10.`,
-            );
-          }
-          return n;
-        },
-      }),
-    ),
-    description:
-      "Keep only the N most recent snapshots of the fork (1-10).",
-  }),
-  keepLastSnapshotsFor: cmd.option({
-    long: "keep-last-snapshots-for",
-    type: cmd.optional(SnapshotExpiration),
-    description:
-      'Expiration applied to kept snapshots. Use "none" or 0 for no expiration. Example: 7d, 30d',
-  }),
-  deleteEvictedSnapshots: cmd.option({
-    long: "delete-evicted-snapshots",
-    type: cmd.optional({
-      ...cmd.oneOf(["true", "false"]),
-      displayName: "true|false",
-    }),
-    description:
-      'When "true" (the default), evicted snapshots are deleted immediately; when "false", they keep the default expiration.',
-  }),
+  ...snapshotRetentionArgs,
   ...networkPolicyArgs,
   scope,
 } as const;
@@ -173,33 +117,11 @@ export const fork = cmd.command({
         })
       : undefined;
 
-    if (
-      keepLastSnapshots === undefined &&
-      (keepLastSnapshotsFor !== undefined ||
-        deleteEvictedSnapshots !== undefined)
-    ) {
-      throw new Error(
-        [
-          "--keep-last-snapshots-for and --delete-evicted-snapshots require --keep-last-snapshots.",
-          `${chalk.bold("hint:")} Pass --keep-last-snapshots <count> to enable the retention policy.`,
-        ].join("\n"),
-      );
-    }
-
-    const keepLastSnapshotsPayload =
-      keepLastSnapshots !== undefined
-        ? {
-            count: keepLastSnapshots,
-            expiration:
-              keepLastSnapshotsFor !== undefined
-                ? ms(keepLastSnapshotsFor)
-                : undefined,
-            deleteEvicted:
-              deleteEvictedSnapshots !== undefined
-                ? deleteEvictedSnapshots === "true"
-                : undefined,
-          }
-        : undefined;
+    const keepLastSnapshotsPayload = buildKeepLastSnapshotsPayload({
+      keepLastSnapshots,
+      keepLastSnapshotsFor,
+      deleteEvictedSnapshots,
+    });
 
     const tagsObj = Object.keys(tags).length > 0 ? tags : undefined;
     const envObj = Object.keys(envVars).length > 0 ? envVars : undefined;
@@ -240,40 +162,12 @@ export const fork = cmd.command({
       );
     }
 
-    const routes = sandbox.routes.filter(
-      (x) => x.port !== sandbox.interactivePort,
-    );
-
     if (!silent) {
-      const teamDisplay = scope.teamSlug ?? scope.team;
-      const projectDisplay = scope.projectSlug ?? scope.project;
-      const hasPorts = routes.length > 0;
-
-      process.stderr.write("✅ Sandbox ");
-      process.stdout.write(chalk.cyan(sandbox.name));
-      process.stderr.write(" forked from " + chalk.cyan(source) + ".\n");
-      process.stderr.write(
-        chalk.dim("   │ ") + "team: " + chalk.cyan(teamDisplay) + "\n",
-      );
-
-      if (hasPorts) {
-        process.stderr.write(
-          chalk.dim("   │ ") + "project: " + chalk.cyan(projectDisplay) + "\n",
-        );
-        process.stderr.write(chalk.dim("   │ ") + "ports:\n");
-        for (let i = 0; i < routes.length; i++) {
-          const route = routes[i];
-          const isLast = i === routes.length - 1;
-          const prefix = isLast ? chalk.dim("   ╰ ") : chalk.dim("   │ ");
-          process.stderr.write(
-            prefix + "• " + route.port + " -> " + chalk.cyan(route.url) + "\n",
-          );
-        }
-      } else {
-        process.stderr.write(
-          chalk.dim("   ╰ ") + "project: " + chalk.cyan(projectDisplay) + "\n",
-        );
-      }
+      printSandboxSummary({
+        sandbox,
+        scope,
+        action: `forked from ${chalk.cyan(source)}`,
+      });
     }
 
     if (connect) {
